abstract out parseFiles
