+++ src/parseInput/parseInput.js
@@ -8,6 +8,7 @@ import { createDelimiter } from './createDelimiter';
 import { refreshPreview } from './refreshPreview';
 import { makePage } from './makePage';
 import { filterColumns } from './filterColumns';
+import { parseFiles } from './parseFiles';
 
 import { flowPreludeFunction } from '../flowPreludeFunction';
 const flowPrelude = flowPreludeFunction();
@@ -116,39 +117,6 @@ export function parseInput() {
       const start = currentPage.index * MaxItemsPerPage;
       return currentPage.columns.slice(start, start + MaxItemsPerPage);
     });
-    const parseFiles = () => {
-      let column;
-      let columnNames;
-      let headerOption;
-      columnNames = (() => {
-        let _i;
-        let _len;
-        const _ref = _columns();
-        const _results = [];
-        for (_i = 0, _len = _ref.length; _i < _len; _i++) {
-          column = _ref[_i];
-          _results.push(column.name());
-        }
-        return _results;
-      })();
-      headerOption = _headerOptions[_headerOption()];
-      if (lodash.every(columnNames, columnName => columnName.trim() === '')) {
-        columnNames = null;
-        headerOption = -1;
-      }
-      const columnTypes = (() => {
-        let _i;
-        let _len;
-        const _ref = _columns();
-        const _results = [];
-        for (_i = 0, _len = _ref.length; _i < _len; _i++) {
-          column = _ref[_i];
-          _results.push(column.type());
-        }
-        return _results;
-      })();
-      return _.insertAndExecuteCell('cs', 'parseFiles\n  ' + _inputKey + ': ' + flowPrelude.stringify(_inputs[_inputKey]) + '\n  destination_frame: ' + flowPrelude.stringify(_destinationKey()) + '\n  parse_type: ' + flowPrelude.stringify(_parseType().type) + '\n  separator: ' + _delimiter().charCode + '\n  number_columns: ' + _columnCount() + '\n  single_quotes: ' + _useSingleQuotes() + '\n  ' + (_canReconfigure() ? 'column_names: ' + flowPrelude.stringify(columnNames) + '\n  ' : '') + (_canReconfigure() ? 'column_types: ' + flowPrelude.stringify(columnTypes) + '\n  ' : '') + 'delete_on_done: ' + _deleteOnDone() + '\n  check_header: ' + headerOption + '\n  chunk_size: ' + _chunkSize()); // eslint-disable-line
-    };
     const _canGoToNextPage = Flow.Dataflow.lift(_activePage, currentPage => (currentPage.index + 1) * MaxItemsPerPage < currentPage.columns.length);
     const _canGoToPreviousPage = Flow.Dataflow.lift(_activePage, currentPage => currentPage.index > 0);
     const goToNextPage = () => {
